% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{Empty}
\alias{Empty}
\title{Google Cloud Logging API Objects 
Writes log entries and manages your logs, log sinks, and logs-based metrics.}
\usage{
Empty()
}
\value{
Empty object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:20:29
filename: /Users/mark/dev/R/autoGoogleAPI/googleloggingv2beta1.auto/R/logging_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Empty Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for `Empty` is empty JSON object `{}`.
}

