% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_functions.R
\name{entries.write}
\alias{entries.write}
\title{Writes log entries to Cloud Logging. All log entries in Cloud Logging are written by this method.}
\usage{
entries.write(WriteLogEntriesRequest)
}
\arguments{
\item{WriteLogEntriesRequest}{The \link{WriteLogEntriesRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/logging.admin
\item https://www.googleapis.com/auth/logging.write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/logging.admin, https://www.googleapis.com/auth/logging.write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/logging/docs/}{Google Documentation}

Other WriteLogEntriesRequest functions: \code{\link{WriteLogEntriesRequest.labels}},
  \code{\link{WriteLogEntriesRequest}}
}

