% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{HttpRequest}
\alias{HttpRequest}
\title{HttpRequest Object}
\usage{
HttpRequest(requestMethod = NULL, requestUrl = NULL, requestSize = NULL,
  status = NULL, responseSize = NULL, userAgent = NULL, remoteIp = NULL,
  referer = NULL, cacheLookup = NULL, cacheHit = NULL,
  cacheValidatedWithOriginServer = NULL, cacheFillBytes = NULL)
}
\arguments{
\item{requestMethod}{The request method}

\item{requestUrl}{The scheme (http, https), the host name, the path and the query portion of the URL that was requested}

\item{requestSize}{The size of the HTTP request message in bytes, including the request headers and the request body}

\item{status}{The response code indicating the status of response}

\item{responseSize}{The size of the HTTP response message sent back to the client, in bytes, including the response headers and the response body}

\item{userAgent}{The user agent sent by the client}

\item{remoteIp}{The IP address (IPv4 or IPv6) of the client that issued the HTTP request}

\item{referer}{The referer URL of the request, as defined in [HTTP/1}

\item{cacheLookup}{Whether or not a cache lookup was attempted}

\item{cacheHit}{Whether or not an entity was served from cache (with or without validation)}

\item{cacheValidatedWithOriginServer}{Whether or not the response was validated with the origin server before being served from cache}

\item{cacheFillBytes}{The number of HTTP response bytes inserted into cache}
}
\value{
HttpRequest object
}
\description{
HttpRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A common proto for logging HTTP requests.
}

