% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{ListLogEntriesRequest}
\alias{ListLogEntriesRequest}
\title{ListLogEntriesRequest Object}
\usage{
ListLogEntriesRequest(projectIds = NULL, filter = NULL, orderBy = NULL,
  pageSize = NULL, pageToken = NULL, partialSuccess = NULL)
}
\arguments{
\item{projectIds}{Required}

\item{filter}{Optional}

\item{orderBy}{Optional}

\item{pageSize}{Optional}

\item{pageToken}{Optional}

\item{partialSuccess}{Optional}
}
\value{
ListLogEntriesRequest object
}
\description{
ListLogEntriesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The parameters to `ListLogEntries`.
}
\seealso{
Other ListLogEntriesRequest functions: \code{\link{entries.list}}
}

