% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{ListLogEntriesResponse.projectIdErrors}
\alias{ListLogEntriesResponse.projectIdErrors}
\title{ListLogEntriesResponse.projectIdErrors Object}
\usage{
ListLogEntriesResponse.projectIdErrors()
}
\value{
ListLogEntriesResponse.projectIdErrors object
}
\description{
ListLogEntriesResponse.projectIdErrors Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
If partial_success is true, contains the project ids that had errors and the associated errors.
}
\seealso{
Other ListLogEntriesResponse functions: \code{\link{ListLogEntriesResponse}}
}

