% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{ListLogEntriesResponse}
\alias{ListLogEntriesResponse}
\title{ListLogEntriesResponse Object}
\usage{
ListLogEntriesResponse(ListLogEntriesResponse.projectIdErrors = NULL,
  entries = NULL, nextPageToken = NULL, projectIdErrors = NULL)
}
\arguments{
\item{ListLogEntriesResponse.projectIdErrors}{The \link{ListLogEntriesResponse.projectIdErrors} object or list of objects}

\item{entries}{A list of log entries}

\item{nextPageToken}{If there are more results than were returned, then `nextPageToken` is included in the response}

\item{projectIdErrors}{If partial_success is true, contains the project ids that had errors and the associated errors}
}
\value{
ListLogEntriesResponse object
}
\description{
ListLogEntriesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Result returned from `ListLogEntries`.
}
\seealso{
Other ListLogEntriesResponse functions: \code{\link{ListLogEntriesResponse.projectIdErrors}}
}

