% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{ListLogMetricsResponse}
\alias{ListLogMetricsResponse}
\title{ListLogMetricsResponse Object}
\usage{
ListLogMetricsResponse(metrics = NULL, nextPageToken = NULL)
}
\arguments{
\item{metrics}{A list of logs-based metrics}

\item{nextPageToken}{If there are more results than were returned, then `nextPageToken` is included in the response}
}
\value{
ListLogMetricsResponse object
}
\description{
ListLogMetricsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Result returned from ListLogMetrics.
}

