% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{ListMonitoredResourceDescriptorsResponse}
\alias{ListMonitoredResourceDescriptorsResponse}
\title{ListMonitoredResourceDescriptorsResponse Object}
\usage{
ListMonitoredResourceDescriptorsResponse(resourceDescriptors = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{resourceDescriptors}{A list of resource descriptors}

\item{nextPageToken}{If there might be more results than those appearing in this response, then nextPageToken is included}
}
\value{
ListMonitoredResourceDescriptorsResponse object
}
\description{
ListMonitoredResourceDescriptorsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Result returned from ListMonitoredResourceDescriptors.
}
