% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{ListSinksResponse}
\alias{ListSinksResponse}
\title{ListSinksResponse Object}
\usage{
ListSinksResponse(sinks = NULL, nextPageToken = NULL)
}
\arguments{
\item{sinks}{A list of sinks}

\item{nextPageToken}{If there are more results than were returned, then `nextPageToken` is included in the response}
}
\value{
ListSinksResponse object
}
\description{
ListSinksResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Result returned from `ListSinks`.
}

