% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{LogEntryOperation}
\alias{LogEntryOperation}
\title{LogEntryOperation Object}
\usage{
LogEntryOperation(id = NULL, producer = NULL, first = NULL, last = NULL)
}
\arguments{
\item{id}{Required}

\item{producer}{Required}

\item{first}{Optional}

\item{last}{Optional}
}
\value{
LogEntryOperation object
}
\description{
LogEntryOperation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Additional information about a potentially long-running operation with which a log entry is associated.
}

