% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{LogMetric}
\alias{LogMetric}
\title{LogMetric Object}
\usage{
LogMetric(name = NULL, description = NULL, filter = NULL)
}
\arguments{
\item{name}{Required}

\item{description}{A description of this metric, which is used in documentation}

\item{filter}{An [advanced logs filter](/logging/docs/view/advanced_filters)}
}
\value{
LogMetric object
}
\description{
LogMetric Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes a logs-based metric. The value of the metric is the number of log entries that match a logs filter.
}
\seealso{
Other LogMetric functions: \code{\link{projects.metrics.create}},
  \code{\link{projects.metrics.update}}
}

