% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{LogSink}
\alias{LogSink}
\title{LogSink Object}
\usage{
LogSink(filter = NULL, destination = NULL, endTime = NULL,
  startTime = NULL, writerIdentity = NULL, outputVersionFormat = NULL,
  name = NULL)
}
\arguments{
\item{filter}{Optional}

\item{destination}{Required}

\item{endTime}{Optional}

\item{startTime}{Optional}

\item{writerIdentity}{Output only}

\item{outputVersionFormat}{Optional}

\item{name}{Required}
}
\value{
LogSink object
}
\description{
LogSink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes a sink used to export log entries to one of the following destinations in any project: a Cloud Storage bucket, a BigQuery dataset, or a Cloud Pub/Sub topic. A logs filter controls which log entries are exported. The sink must be created within a project, organization, billing account, or folder.
}
\seealso{
Other LogSink functions: \code{\link{projects.sinks.create}},
  \code{\link{projects.sinks.update}}
}
