% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{MonitoredResource}
\alias{MonitoredResource}
\title{MonitoredResource Object}
\usage{
MonitoredResource(MonitoredResource.labels = NULL, type = NULL,
  labels = NULL)
}
\arguments{
\item{MonitoredResource.labels}{The \link{MonitoredResource.labels} object or list of objects}

\item{type}{Required}

\item{labels}{Required}
}
\value{
MonitoredResource object
}
\description{
MonitoredResource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An object representing a resource that can be used for monitoring, logging, billing, or other purposes. Examples include virtual machine instances, databases, and storage devices such as disks. The `type` field identifies a MonitoredResourceDescriptor object that describes the resource's schema. Information in the `labels` field identifies the actual resource and its attributes according to the schema. For example, a particular Compute Engine VM instance could be represented by the following object, because the MonitoredResourceDescriptor for `'gce_instance'` has labels `'instance_id'` and `'zone'`: { 'type': 'gce_instance', 'labels': { 'instance_id': 'my-instance', 'zone': 'us-central1-a' }}
}
\seealso{
Other MonitoredResource functions: \code{\link{MonitoredResource.labels}}
}

