% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_functions.R
\name{projects.sinks.update}
\alias{projects.sinks.update}
\title{Creates or updates a sink.}
\usage{
projects.sinks.update(LogSink, sinkName)
}
\arguments{
\item{LogSink}{The \link{LogSink} object to pass to this method}

\item{sinkName}{The resource name of the sink to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/logging.admin
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/logging.admin)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/logging/docs/}{Google Documentation}

Other LogSink functions: \code{\link{LogSink}},
  \code{\link{projects.sinks.create}}
}

