% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{RequestLog}
\alias{RequestLog}
\title{RequestLog Object}
\usage{
RequestLog(appId = NULL, moduleId = NULL, versionId = NULL,
  requestId = NULL, ip = NULL, startTime = NULL, endTime = NULL,
  latency = NULL, megaCycles = NULL, method = NULL, resource = NULL,
  httpVersion = NULL, status = NULL, responseSize = NULL,
  referrer = NULL, userAgent = NULL, nickname = NULL,
  urlMapEntry = NULL, host = NULL, cost = NULL, taskQueueName = NULL,
  taskName = NULL, wasLoadingRequest = NULL, pendingTime = NULL,
  instanceIndex = NULL, finished = NULL, first = NULL,
  instanceId = NULL, line = NULL, appEngineRelease = NULL,
  traceId = NULL, sourceReference = NULL)
}
\arguments{
\item{appId}{Application that handled this request}

\item{moduleId}{Module of the application that handled this request}

\item{versionId}{Version of the application that handled this request}

\item{requestId}{Globally unique identifier for a request, which is based on the request start time}

\item{ip}{Origin IP address}

\item{startTime}{Time when the request started}

\item{endTime}{Time when the request finished}

\item{latency}{Latency of the request}

\item{megaCycles}{Number of CPU megacycles used to process request}

\item{method}{Request method}

\item{resource}{Contains the path and query portion of the URL that was requested}

\item{httpVersion}{HTTP version of request}

\item{status}{HTTP response status code}

\item{responseSize}{Size in bytes sent back to client by request}

\item{referrer}{Referrer URL of request}

\item{userAgent}{User agent that made the request}

\item{nickname}{The logged-in user who made the request}

\item{urlMapEntry}{File or class that handled the request}

\item{host}{Internet host and port number of the resource being requested}

\item{cost}{An indication of the relative cost of serving this request}

\item{taskQueueName}{Queue name of the request, in the case of an offline request}

\item{taskName}{Task name of the request, in the case of an offline request}

\item{wasLoadingRequest}{Whether this was a loading request for the instance}

\item{pendingTime}{Time this request spent in the pending request queue}

\item{instanceIndex}{If the instance processing this request belongs to a manually scaled module, then this is the 0-based index of the instance}

\item{finished}{Whether this request is finished or active}

\item{first}{Whether this is the first RequestLog entry for this request}

\item{instanceId}{An identifier for the instance that handled the request}

\item{line}{A list of log lines emitted by the application while serving this request}

\item{appEngineRelease}{App Engine release version}

\item{traceId}{Cloud Trace identifier for this request}

\item{sourceReference}{Source code for the application that handled this request}
}
\value{
RequestLog object
}
\description{
RequestLog Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Complete log information about a single HTTP request to an App Engine application.
}

