% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{SourceReference}
\alias{SourceReference}
\title{SourceReference Object}
\usage{
SourceReference(revisionId = NULL, repository = NULL)
}
\arguments{
\item{revisionId}{The canonical and persistent identifier of the deployed revision}

\item{repository}{Optional}
}
\value{
SourceReference object
}
\description{
SourceReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A reference to a particular snapshot of the source tree used to build and deploy an application.
}
