% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{WriteLogEntriesRequest.labels}
\alias{WriteLogEntriesRequest.labels}
\title{WriteLogEntriesRequest.labels Object}
\usage{
WriteLogEntriesRequest.labels()
}
\value{
WriteLogEntriesRequest.labels object
}
\description{
WriteLogEntriesRequest.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional. User-defined `key:value` items that are added to the `labels` field of each log entry in `entries`, except when a log entry specifies its own `key:value` item with the same key. Example: `{ 'size': 'large', 'color':'red' }`
}
\seealso{
Other WriteLogEntriesRequest functions: \code{\link{WriteLogEntriesRequest}},
  \code{\link{entries.write}}
}

