% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{WriteLogEntriesResponse}
\alias{WriteLogEntriesResponse}
\title{WriteLogEntriesResponse Object}
\usage{
WriteLogEntriesResponse()
}
\value{
WriteLogEntriesResponse object
}
\description{
WriteLogEntriesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Result returned from WriteLogEntries. empty
}

