% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{ListLogEntriesRequest}
\alias{ListLogEntriesRequest}
\title{ListLogEntriesRequest Object}
\usage{
ListLogEntriesRequest(orderBy = NULL, resourceNames = NULL,
  projectIds = NULL, filter = NULL, pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{orderBy}{Optional}

\item{resourceNames}{Required}

\item{projectIds}{Deprecated}

\item{filter}{Optional}

\item{pageToken}{Optional}

\item{pageSize}{Optional}
}
\value{
ListLogEntriesRequest object
}
\description{
ListLogEntriesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The parameters to ListLogEntries.
}
\seealso{
Other ListLogEntriesRequest functions: \code{\link{entries.list}}
}
