% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{LogEntry}
\alias{LogEntry}
\title{LogEntry Object}
\usage{
LogEntry(LogEntry.protoPayload = NULL, LogEntry.jsonPayload = NULL,
  LogEntry.labels = NULL, logName = NULL, resource = NULL,
  protoPayload = NULL, textPayload = NULL, jsonPayload = NULL,
  timestamp = NULL, severity = NULL, insertId = NULL,
  httpRequest = NULL, labels = NULL, operation = NULL)
}
\arguments{
\item{LogEntry.protoPayload}{The \link{LogEntry.protoPayload} object or list of objects}

\item{LogEntry.jsonPayload}{The \link{LogEntry.jsonPayload} object or list of objects}

\item{LogEntry.labels}{The \link{LogEntry.labels} object or list of objects}

\item{logName}{Required}

\item{resource}{Required}

\item{protoPayload}{The log entry payload, represented as a protocol buffer}

\item{textPayload}{The log entry payload, represented as a Unicode string (UTF-8)}

\item{jsonPayload}{The log entry payload, represented as a structure that is expressed as a JSON object}

\item{timestamp}{Optional}

\item{severity}{Optional}

\item{insertId}{Optional}

\item{httpRequest}{Optional}

\item{labels}{Optional}

\item{operation}{Optional}
}
\value{
LogEntry object
}
\description{
LogEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An individual entry in a log.
}
\seealso{
Other LogEntry functions: \code{\link{LogEntry.jsonPayload}},
  \code{\link{LogEntry.labels}},
  \code{\link{LogEntry.protoPayload}}
}

