% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{LogLine}
\alias{LogLine}
\title{Stackdriver Logging API Objects 
Writes log entries and manages your Stackdriver Logging configuration.}
\usage{
LogLine(time = NULL, severity = NULL, logMessage = NULL,
  sourceLocation = NULL)
}
\arguments{
\item{time}{Approximate time when this log entry was made}

\item{severity}{Severity of this log entry}

\item{logMessage}{App-provided log message}

\item{sourceLocation}{Where in the source code this log message was written}
}
\value{
LogLine object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:01:25
filename: /Users/mark/dev/R/autoGoogleAPI/googleloggingv2beta1.auto/R/logging_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
LogLine Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Application log line emitted while processing a request.
}
