% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{RequestLog}
\alias{RequestLog}
\title{RequestLog Object}
\usage{
RequestLog(userAgent = NULL, wasLoadingRequest = NULL,
  sourceReference = NULL, responseSize = NULL, traceId = NULL,
  line = NULL, referrer = NULL, taskQueueName = NULL, requestId = NULL,
  nickname = NULL, pendingTime = NULL, resource = NULL, status = NULL,
  taskName = NULL, urlMapEntry = NULL, instanceIndex = NULL,
  finished = NULL, host = NULL, httpVersion = NULL, startTime = NULL,
  latency = NULL, ip = NULL, appId = NULL, appEngineRelease = NULL,
  method = NULL, cost = NULL, instanceId = NULL, megaCycles = NULL,
  first = NULL, versionId = NULL, moduleId = NULL, endTime = NULL)
}
\arguments{
\item{userAgent}{User agent that made the request}

\item{wasLoadingRequest}{Whether this was a loading request for the instance}

\item{sourceReference}{Source code for the application that handled this request}

\item{responseSize}{Size in bytes sent back to client by request}

\item{traceId}{Stackdriver Trace identifier for this request}

\item{line}{A list of log lines emitted by the application while serving this request}

\item{referrer}{Referrer URL of request}

\item{taskQueueName}{Queue name of the request, in the case of an offline request}

\item{requestId}{Globally unique identifier for a request, which is based on the request start time}

\item{nickname}{The logged-in user who made the request}

\item{pendingTime}{Time this request spent in the pending request queue}

\item{resource}{Contains the path and query portion of the URL that was requested}

\item{status}{HTTP response status code}

\item{taskName}{Task name of the request, in the case of an offline request}

\item{urlMapEntry}{File or class that handled the request}

\item{instanceIndex}{If the instance processing this request belongs to a manually scaled module, then this is the 0-based index of the instance}

\item{finished}{Whether this request is finished or active}

\item{host}{Internet host and port number of the resource being requested}

\item{httpVersion}{HTTP version of request}

\item{startTime}{Time when the request started}

\item{latency}{Latency of the request}

\item{ip}{Origin IP address}

\item{appId}{Application that handled this request}

\item{appEngineRelease}{App Engine release version}

\item{method}{Request method}

\item{cost}{An indication of the relative cost of serving this request}

\item{instanceId}{An identifier for the instance that handled the request}

\item{megaCycles}{Number of CPU megacycles used to process request}

\item{first}{Whether this is the first RequestLog entry for this request}

\item{versionId}{Version of the application that handled this request}

\item{moduleId}{Module of the application that handled this request}

\item{endTime}{Time when the request finished}
}
\value{
RequestLog object
}
\description{
RequestLog Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Complete log information about a single HTTP request to an App Engine application.
}
