% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{SourceLocation}
\alias{SourceLocation}
\title{SourceLocation Object}
\usage{
SourceLocation(file = NULL, line = NULL, functionName = NULL)
}
\arguments{
\item{file}{Source file name}

\item{line}{Line within the source file}

\item{functionName}{Human-readable name of the function or method being invoked, with optional context such as the class or package name}
}
\value{
SourceLocation object
}
\description{
SourceLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies a location in a source code file.
}

