% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_objects.R
\name{WriteLogEntriesRequest}
\alias{WriteLogEntriesRequest}
\title{WriteLogEntriesRequest Object}
\usage{
WriteLogEntriesRequest(WriteLogEntriesRequest.labels = NULL, logName = NULL,
  resource = NULL, labels = NULL, entries = NULL, partialSuccess = NULL)
}
\arguments{
\item{WriteLogEntriesRequest.labels}{The \link{WriteLogEntriesRequest.labels} object or list of objects}

\item{logName}{Optional}

\item{resource}{Optional}

\item{labels}{Optional}

\item{entries}{Required}

\item{partialSuccess}{Optional}
}
\value{
WriteLogEntriesRequest object
}
\description{
WriteLogEntriesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The parameters to WriteLogEntries.
}
\seealso{
Other WriteLogEntriesRequest functions: \code{\link{WriteLogEntriesRequest.labels}},
  \code{\link{entries.write}}
}

